Adapted Mathis's new Int stats and added Amelioration

+++ TinyLittleWarriors/Assets/Script/Stats/CharacterStats.cs
@@ -62,7 +62,7 @@ public class CharacterStats : MonoBehaviour
     }
 
     // Getters and Setters for the Defense
-    public float getDefense()
+    public int getDefense()
     {
         return defense;
     }
